flask_project/requirements.txt
black==23.10.0
Flask==3.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
SQLAlchemy==2.0.22
